stock_update/requirements.txt
pandas
pandas_gbq
protobuf
python-dotenv
Requests
